orig-requirements.txt
anyio==4.4.0
blinker==1.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
filelock==3.14.0
Flask==3.0.3
fsspec==2024.3.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.23.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
loguru==0.7.2
lxml==5.2.1
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
ollama==0.2.1
packaging==24.0
pandas==2.2.2
pillow==10.3.0
plum-dispatch==1.7.4
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
python-docx==1.1.2
pytz==2024.1
PyYAML==6.0.1
regex==2024.4.28
requests==2.32.0
safetensors==0.4.3
scikit-learn==1.4.2
scipy==1.13.0
six==1.16.0
sniffio==1.3.1
Spire.Doc==12.4.0
sympy==1.12
threadpoolctl==3.5.0
tokenizers==0.19.1
torch==2.3.0
tqdm==4.66.4
triton==2.3.0
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
Werkzeug==3.0.3